Move refinement-specific agent key and update imports
