+++ indra_world/assembly/operations.py
@@ -13,8 +13,7 @@ __all__ = ['get_expanded_events_influences', 'remove_namespaces',
            'make_compositional_refinement_filter',
            'make_default_compositional_refinement_filter',
            'CompositionalRefinementFilter', 'get_relevants_for_stmt',
-           'get_agent_key', 'listify', 'comp_ontology', 'comp_ontology_url',
-           'merge_deltas']
+           'listify', 'merge_deltas']
 import os
 import yaml
 import copy
@@ -28,7 +27,6 @@ from indra.sources.eidos.client import reground_texts
 from indra.pipeline import register_pipeline
 from indra.statements import Influence, Association, Event
 from indra.statements.concept import get_sorted_compositional_groundings
-from indra_world.ontology.ontology import WorldOntology
 
 from .matches import *
 from .refinement import *
@@ -37,12 +35,6 @@ logger = logging.getLogger(__name__)
 
 
 register_pipeline(datetime)
-
-comp_onto_branch = '4531c084d3b902f04605c11396a25db4fff16573'
-comp_ontology_url = 'https://raw.githubusercontent.com/WorldModelers/'\
-                    'Ontologies/%s/CompositionalOntology_v2.1_metadata.yml' % \
-    comp_onto_branch
-comp_ontology = WorldOntology(comp_ontology_url)
 
 
 @register_pipeline
@@ -467,40 +459,6 @@ def merge_deltas(stmts_in):
             attr.adjectives = all_adjectives
         stmts_out.append(stmt)
     return stmts_out
-
-
-def get_agent_key(agent, comp_idx):
-    """Return a key for an Agent for use in refinement finding.
-
-    Parameters
-    ----------
-    agent : indra.statements.Agent or None
-         An INDRA Agent whose key should be returned.
-
-    Returns
-    -------
-    tuple or None
-        The key that maps the given agent to the ontology, with special
-        handling for ungrounded and None Agents.
-    """
-    # Possibilities are as follows:
-    # Case 1: no grounding, in which case we use the agent name as a theme
-    # grounding. If we are looking at another component, we return None.
-    # Case 2: There is a WM compositional grounding in which case we return
-    # the specific entry in the compositional tuple if available, or None if
-    # not.
-    if isinstance(agent, Event):
-        agent = agent.concept
-    gr = agent.get_grounding(ns_order=['WM'])
-    if gr[0] is None:
-        if comp_idx == 0:
-            agent_key = ('NAME', agent.name)
-        else:
-            agent_key = None
-    else:
-        comp_gr = gr[1][comp_idx]
-        agent_key = ('WM', comp_gr) if comp_gr else None
-    return agent_key
 
 
 @register_pipeline
